+++ src/run.rs
@@ -1451,6 +1451,8 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                 Key::Char('u') => {
                     let mani_len = state.operations.op_list.len();
                     if mani_len < state.operations.pos + 1 {
+                        print_info("No operations left.", y);
+                        screen.flush()?;
                         continue;
                     }
                     if let Some(op) = state
@@ -1485,6 +1487,8 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                     let mani_len = state.operations.op_list.len();
                     if mani_len == 0 || state.operations.pos == 0 || mani_len < state.operations.pos
                     {
+                        print_info("No operations left.", y);
+                        screen.flush()?;
                         continue;
                     }
                     if let Some(op) = state